requirements.txt
absl-py==0.1.10
autopep8==1.3.4
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cymem==1.31.2
cytoolz==0.8.2
decorator==4.2.1
dill==0.2.7.1
ftfy==4.4.3
futures==3.1.1
grpcio==1.9.1
html5lib==0.9999999
idna==2.6
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.11
msgpack-numpy==0.4.1
msgpack-python==0.5.1
murmurhash==0.28.0
nltk==3.2.5
numpy==1.14.1
parso==0.1.1
pathlib==1.0.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pycodestyle==2.3.1
Pygments==2.2.0
PyYAML==3.12
regex==2017.4.5
requests==2.18.4
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.19.5
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
